Create primitive login system
Login system is composed by Authentication and Permission library.
Authentication for doing the authentication system and Permission for
checking the user state.

NOTE: Session files is stored in sessions directory that web readable.
This is a bad practice and still thinking to relocate them. It is just
termporary place and protected by htaccess file.