Code documentation
Add documentation comment inside the code it self for ensuring
future project development progress. As a (good) side effect,
another programmer can join in to improve this project.

Creating readme file as a reminder the root of this project.
Also, make sure people know what Dashvolt is. License file
is created to specify and clarify which license is aplied
to my own work, which license is from another project
(because I use existing framework).